[Fix] : 뉴스 상세페이지 스크랩, 신고 기능 수정
